AGENT-456 Add debugging information for failed connectivity checks
Move status view updates into a function so that it can be reused.

If release image cannot be pulled, additional checks are performed:
* The host where the release image is located is pinged
* The host's domain name is looked up

The connectivity check results are stored in a struct so that
it can be used during application started to access if network
configuration is required. If network configuration is required
the results can then be immediately displayed in the status view
without running another set of checks.

Navigation on the regnode form has been changed. Left and right
arrows moves between buttons. Tab keys move between the pane
containing the buttons and the status view pane.

Signed-off-by: Richard Su <rwsu@redhat.com>