Enumerate the number of errors
Tview appears to not support scrollbars in their windows. When
there is a large number of errors, the text view could automatically
scrolls to the end. It may not be obvious there is more text to
view.

Enumerating the number of errors would indicate to the users there
are additional details to see, which may prompt them to switch
to the status window and scroll.

A complicating factor is, it is not obvious how to switch to the
status window and which keys allow scrolling up and down.