AGENT-464: Adds form to agent-tui to enter rendezvous IP address
This change supports a new user scenario whereby the rendezvous
IP address may not be written into the agent ISO. In this case,
the agent-tui should detect if the NODE_ZERO_IP is specified
in /etc/assisted/rendezvous-host.env. If it is unspecified, the
tui should display a form to allow the user to enter the
rendezvous host IP.

There will be a supporting change in the installer repo to
pass the NODE_ZERO_IP from /etc/assisted/rendezvous-host.env
to the agent-interactive-console.service which executes
agent-tui during boot.