Refactor UI for Rendezvous node IP entry
Added confirmation screens after the Rendezvous node IP is saved.

Instead of a checkbox to "randomly" choose of the Node's IP as
the Rendezvous node IP, a list of all of this node's IP addresses
is presented which allows the user to choose the exact one they
want.

Added additional descriptions on the form to make the process
more understandable.