OCPBUGS-56223: Fix nmtui redirection after quit
After quiting nmtui in the install-interactive-disconnected workflow
the user should stay in the rendezvous IP selection screens and
not shown the release image checks page.

The bug was caused by
https://github.com/openshift/agent-installer-utils/pull/73
which removed the original redirect back to the rendezvous IP pages
to fix an issue with the bad_dns e2e test.