OCPBUGS-56147: Add operator CRs to ignition in OVE build script
Some operator CRs need to be applied after installation is complete,
this will be done by the appliance in
https://github.com/openshift/appliance/pull/398.
This PR adds the CRs to ignition so that will be populated
to /etc/assisted/extra-manifests and be available for the appliance.

For the kubedescheduler operator the output of the CR matches
what is in https://issues.redhat.com/browse/MGMT-20588.