NO-ISSUE: Enable 'q' and 'c' hotkeys for quit and configure
Entering TAB or LEFT keys after ESC is not reliable and sometimes
are not registered through "virsh send-key" leaving the bad_dns
test unable to quit agent-tui. Using TAB or LEFT makes exiting
out of agent-tui a navigational exercise.

To provide a simplier way to exit agent-tui, the checks page has
been updated to recognize the letter 'q' as an alternative means
to exit the agent-tui.

To make the interface consistent, typing 'c' will start the
nmtui from the checks page.