AGENT-1193: Custom openshift-install to generate unconfigued-ign
The openshift appliance build command has a flag called

--debug-base-ignition

when provided, allows using a custom openshift-install binary to
generate the unconfigured-ignition instead of the binary downloaded
from the release.

When the flag is provided, openshift appliance looks for the custom
openshift-install binary in its assets directory.

The build script has been updated to detect if a custom
openshift-install binary has been provided in the assets
directory and if present, adds the --debug-base-ignition flag.

One use case would be for dev-scripts. When the configuration
variable KNI_INSTALL_FROM_GIT is set to true, dev-scripts builds
the openshift-install binary from source and uses it to generate
the agent ISO. We would like the appliance to also use the same
openshift-install binary that was generated from source to build
its images. A supporting PR in dev-scripts will copy the
openshift-install binary to the appliance's asset directory
when KNI_INSTALL_FROM_GIT=true. In combination with this PR, it
will activate the --debug-base-ignition flag.