OCPBUGS-61956: Disable Konflux PR go.mod PR updates
Disable the Konflux Renovate/MintMaker PR updates of go.mod
dependencies.  Changes to files in .tekton will still receive
PRs, for example for updates to Konflux tasks.

In addition the DependencyDashboard is enabled which will allow
Renovate to show recommended dependency updates under Issues,
but not actually create the PR.